requirements.txt
fastapi
uvicorn
python-dotenv
motor
aiohttp
requests
pydantic
msal
google-auth-oauthlib
google-api-python-client
httpx
